requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.
python-docx>=0.8.6
PyYAML>=3.1.0 # MIT
jsonschema # MIT
